requirements.txt
# SmartBot Phase 2
# Pin exact versions with: pip freeze > requirements_lock.txt
opencv-python>=4.8.0
ultralytics>=8.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
